Minimum to use <AuthorizeView> Component Using the AuthorizeView component to show or hide content based on the user's authentication status.
Adding <AuthorizeView> will fail at runtime unless the
AuthorizeRouteView is in Routes.razor.

An AuthenticationStateProvider is required, so the
ServerAuthenticationStateProvider is added as a Scoped service.

Finally, need to AddAuthorization to services in Program.cs.

Once that is done the app should open up to the Home page and show
that the app is trying to obtain and display some of the user's
authentication information will work.